switched to using css for image maps
Switched to using css alongside html image maps to make a hoverable GUI for the US map. A few states have been created that will serve as the hover layout when the mouse moves over them. More still need to be made. Worked on JS code to implement functionality for when the user interacts with the map, specifically only showing hovered states.