Finished retrieval script to retrieve api data
The script used to continuously retrieve and store Accuweather api data for the Waccurate app is finished. It retrieves statistics from the Accuweather api and stores it in a file. All that needs to be done is to continuously scrape data from Accuweather so that the Waccurate app has enough data to make conclusions about how accurate Accuweather really is.