Made api data retrieval more automated and consistent
Less parameters are needed to retreive api data from the Accuweather API in the retrieval.js script. Simplified code in some places. Made it so that the retrieval.js script accounts for two different types of data from the accuweather API: historical data and twelve hour forecast data. For homepage.html, made it so that almost all states on the map of the United States are clickable.