Changed layout for input widgets on the homepage and added more functionality to data_retrieval.js
Scrapped the use of Bootstrap purely for layouts and switched to css grids for arranging input widgets in homepage.html. Made dropdown widgets responsive to user interaction, specifically clicking on dropdown items. Added functionality to retrieval.js so that API data for 12 hour forecasts automatically get appended to other 12 hour forecast json files when the time is 8 or 9 in the morning (arbitrary; this time is convenient for me to run the program and collect data).