Added two functions to database_manager and slightly altered the layout of the waccurate homepage input panel
Added get_temp_stat and get_precip_stat methods to database_manager which both caclulate the accuracy of Accuweather's prediction for a given hour. The hour that is specified by the user is in relation to 9 am. Modified the layout of the input panel on the Waccurate homepage so that there is only one 'ok' button to be clicked once the user has entered all options that they choose