Fixed a few bugs with getting global weather statistic data from data_manager and implemented an API to send that data over the server
Fixed a bug in data_manager.js where an error would be thrown even though there seemed to be nothing wrong with the actual logic. It turned out one of the 12hr data files had the wrong filename, which crashed the program. So far only tested obtaining data for the temperature statistic. Moved the node index file to the data_retrieval folder because otherwise the relative paths for the data folders would not work. A solution needs to be found for this so that index.js is not in the data_retrieval folder. For the index.js file, added routes, including a static path so that when the route containing homepage.html is executed, its respective css and js files get rendered.