fixed and created functions in data_manager that return specific weather statistic data
Fixed and created functions in data_manager that return specific weather statistic data. A function was created to return the percentage accuracy of temperature for every state given a certain amount of hours. The function to return the percentage accuracy of precipitation  given a certain amount of hours was fixed as it returned an error.