Ensured that basic input and output work on the homepage.
The output box is revealed right below the map only when the user has typed in appropriate input. It displays the accuracy has an integer percentage and the input that the user specified. Changed a few element names in homepage.html to better represent what they do. In homepage.js inplemented functionality for user input to be parsed and displayed onto homepage.html. Altered the imageMapResizer.js so that the interactive map on homepage.html can scale in proportion to its image map, but is still not functioning correctly due to individual state image maps not doing the same.